поч классы фигур в файле с классом точек по другим файлам раскидал
